fix(lsp.kotlin_language_server): set proper `storagePath` in init opts
The config from nvim-lspconfig sets wrong `storagePath` which can be
`nil` sometimes. This results in `init_options` being an empty table,
which is then encoded as an array in JSON, crashing the server in the
end. This commit sets initial `storagePath` to an empty string to
prevent `init_options` being encoded as an array and later sets the
proper value in `before_init` according to the `root_dir` of the lsp
client.